+++ src/azlassets/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "3.4.2" # x-release-please-version
+__version__ = "3.5.0" # x-release-please-version